savelogs.sh: Only link files that exist
Sometimes the build may fail too early, for example when the lookaside
cache isn't available.  In those situations, not all log files are
produced.  Let's not fail on that.

+++ misc/savelogs.sh
@@ -8,6 +8,6 @@ cd ~/results/$build/results
 cp -R repodata ~/logs/$build/
 for pkg in $(grep -Fl failed *-status.log|sed 's/-status\.log//'); do
 	for log in build mock-stderr mock-stdout root state status; do
-		ln $pkg-$log.log ~/logs/$build/
+		[ -f $pkg-$log.log ] && ln $pkg-$log.log ~/logs/$build/
 	done
 done
